:truck: chore: rename resources files and add cards-navigation
