+++ src/resources/index.ts
@@ -1,1 +1,2 @@
-export { default as dashboardNavigations } from './dashboardNavigation';
+export { default as dashboardNavigations } from './sideBarNavigation';
+export { default as cardsdNavigation } from './cardsNavigations';
